Almost done with Chapter4

+++ AcceleratedCpp/Chapter4/chap44.cpp
@@ -14,7 +14,7 @@ int main()
 {
 	double logvar = log2(1000);
 	cout << logvar << endl;
-	for(int i = 0; i < 1000; i++)
+	for(double i = 0; i < 1000; i++)
 	{
 		cout <<setw(log2(i)) << i << setw(log2(i*i)) << i*i << endl;
 	}
